+++ 4-tools/node-typescript-compiler/src/find-tsc.js
@@ -5,26 +5,29 @@ const path_exists = require('path-exists')
 const os = require('os')
 
 const { LIB } = require('./consts')
+const { display_banner_if_1st_output } = require('./logger')
 
+const EXECUTABLE = 'tsc'
 
 function find_tsc() {
+
 	// obvious candidate from sibling module,
 	// but won't work if symlinked, with npm link for ex. or with npm-pkgr
-	const candidate_from_sibling_module = path.join(__dirname, '../typescript/bin/tsc')
+	const candidate_from_sibling_module = path.join(__dirname, `../typescript/bin/${EXECUTABLE}`)
 
 	// second try: should work even if module is symlinked
-	const candidate_from_caller_node_module = path.join(process.cwd(), 'node_modules/typescript/bin/tsc')
+	const candidate_from_caller_node_module = path.join(process.cwd(), `node_modules/typescript/bin/${EXECUTABLE}`)
 
 	// 3rd
 	let candidate_from_require = null
 	try {
 		candidate_from_require = path.dirname(require.resolve('typescript'))
-		candidate_from_require = path.join(candidate_from_require, '..', 'bin/tsc')
+		candidate_from_require = path.join(candidate_from_require, '..', `bin/${EXECUTABLE}`)
 	}
 	catch(err) { /* not found, ignore */ }
 
 	// last try: fallbacking to an eventual global typescript module
-	const candidate_from_global = path.join(os.homedir(), '.nvm/versions/node/', process.version, 'bin/tsc')
+	const candidate_from_global = path.join(os.homedir(), '.nvm/versions/node/', process.version, `bin/${EXECUTABLE}`)
 
 
 	function candidate_if_exists(candidate) {
@@ -40,16 +43,18 @@ function find_tsc() {
 		.catch(() => candidate_if_exists(candidate_from_require))
 		.catch(() => candidate_if_exists(candidate_from_global))
 		.catch(() => {
-			console.error(`${LIB}: Couldn’t find a typescript compiler in any expected locations. Unsuccessfully tested locations, by priority:
-- ${candidate_from_sibling_module}
-- ${candidate_from_caller_node_module}
-- ${candidate_from_require} (from require('typescript'))
-- ${candidate_from_global}
-			`)
-			throw new Error('Couldn’t find a typescript compiler in any expected locations!')
+			display_banner_if_1st_output()
+			console.error(`${LIB}: ❌ Couldn’t find a typescript compiler ("${EXECUTABLE}") in any expected locations. Unsuccessfully tested locations, by priority:
+- ${candidate_from_sibling_module} ❌ not found
+- ${candidate_from_caller_node_module} ❌ not found
+- ${candidate_from_require} (from require('typescript')) ❌ not found
+- ${candidate_from_global} ❌ not found
+`)
+			throw new Error(`Couldn’t find the "${EXECUTABLE}" typescript compiler in any expected locations!`)
 		})
 }
 
 module.exports = {
+	EXECUTABLE,
 	find_tsc,
 }